backend/requirements.txt
# =========================
# Core FastAPI
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# =========================
# Database
# =========================
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# =========================
# Image Processing
# =========================
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.6.0.66
scikit-image==0.22.0

# =========================
# AI Vision
# =========================
anthropic==0.39.0

# =========================
# OCR
# =========================
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.7.0.3

# =========================
# PaddleOCR Dependencies
# =========================
paddlepaddle==2.6.2
imgaug==0.4.0
pyclipper==1.3.0.post5
shapely==2.0.3
lmdb==1.5.1
rapidfuzz==3.6.1
visualdl==2.5.3
fire==0.6.0
fonttools==4.49.0
attrdict==2.0.1
Cython==3.0.10

# =========================
# PDF
# =========================
pdfplumber==0.10.3
PyPDF2==3.0.1
PyMuPDF==1.24.1
pdf2image==1.17.0

# =========================
# Data Processing
# =========================
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# =========================
# NLP
# =========================
pyspellchecker==0.8.1

# =========================
# Security
# =========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# =========================
# Utilities
# =========================
python-dateutil==2.9.0.post0
tqdm==4.66.2
PyYAML==6.0.1
protobuf==3.20.2
requests==2.31.0

# =========================
# Testing
# =========================
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0

# =========================
# Biometrics
# =========================
cloudinary==1.36.0
face-recognition==1.3.0
dlib==19.24.6
pyserial==3.5
